pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "po-leu"
version = "0.1.0"
description = "Perceived-Outcome, LLM-generated, Embedding-based Utility (PO-LEU)"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.1",
    "numpy>=1.24",
    "pyyaml>=6.0",
    "sentence-transformers>=2.7",
    "tqdm>=4.66",
    "pandas>=2.0",
    "scikit-learn>=1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-xdist>=3.5",
]
llm = [
    "anthropic>=0.25",
    "python-dotenv>=1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-q"
